+++ src/Silex/ControllerCollection.php
@@ -50,7 +50,7 @@ class ControllerCollection
     public function match($pattern, $to)
     {
         $route = clone $this->defaultRoute;
-        $route->setPattern($pattern);
+        $route->setPath($pattern);
         $route->setDefault('_controller', $to);
 
         $this->controllers[] = $controller = new Controller($route);
@@ -147,6 +147,8 @@ class ControllerCollection
             $routes->add($name, $controller->getRoute());
             $controller->freeze();
         }
+
+        $routes->addPrefix($prefix);
 
         $this->controllers = array();