+++ src/Silex/EventListener/LocaleListener.php
@@ -13,6 +13,7 @@ namespace Silex\EventListener;
 
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\EventListener\LocaleListener as BaseLocaleListener;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Routing\RequestContextAwareInterface;
 use Silex\Application;
 
@@ -25,9 +26,9 @@ class LocaleListener extends BaseLocaleListener
 {
     protected $app;
 
-    public function __construct(Application $app, RequestContextAwareInterface $router = null)
+    public function __construct(Application $app, RequestContextAwareInterface $router = null, RequestStack $requestStack = null)
     {
-        parent::__construct($app['locale'], $router);
+        parent::__construct($app['locale'], $router, $requestStack);
 
         $this->app = $app;
     }