+++ src/Silex/Util/Compiler.php
@@ -12,7 +12,6 @@
 namespace Silex\Util;
 
 use Symfony\Component\Finder\Finder;
-use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\Process\Process;
 
 /**
@@ -88,7 +87,7 @@ class Compiler
         unset($phar);
     }
 
-    protected function addFile($phar, $file, $strip = true)
+    protected function addFile(\Phar $phar, \SplFileInfo $file, $strip = true)
     {
         $path = str_replace(dirname(dirname(dirname(__DIR__))).DIRECTORY_SEPARATOR, '', $file->getRealPath());