+++ tests/Silex/Tests/Provider/MonologServiceProviderTest.php
@@ -15,6 +15,7 @@ use Monolog\Handler\TestHandler;
 use Monolog\Logger;
 use Silex\Application;
 use Silex\Provider\MonologServiceProvider;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -24,13 +25,6 @@ use Symfony\Component\HttpFoundation\Request;
  */
 class MonologServiceProviderTest extends \PHPUnit_Framework_TestCase
 {
-    public function setUp()
-    {
-        if (!is_dir(__DIR__.'/../../../../vendor/monolog/monolog/src')) {
-            $this->markTestSkipped('Monolog dependency was not installed.');
-        }
-    }
-
     public function testRequestLogging()
     {
         $app = $this->getApplication();
@@ -99,6 +93,47 @@ class MonologServiceProviderTest extends \PHPUnit_Framework_TestCase
 
         $pattern = "#RuntimeException: very bad error \(uncaught exception\) at .* line \d+#";
         $this->assertMatchingRecord($pattern, Logger::CRITICAL, $app['monolog.handler']);
+    }
+
+    public function testRedirectLogging()
+    {
+        $app = $this->getApplication();
+
+        $app->get('/foo', function () use ($app) {
+            return new RedirectResponse("/bar", 302);
+        });
+
+        $this->assertFalse($app['monolog.handler']->hasInfoRecords());
+
+        $request = Request::create('/foo');
+        $app->handle($request);
+
+        $this->assertTrue($app['monolog.handler']->hasInfo('< 302 /bar'));
+    }
+
+    public function testErrorLoggingGivesWayToSecurityExceptionHandling()
+    {
+        $app = $this->getApplication();
+        $app['monolog.level'] = Logger::ERROR;
+
+        $app->register(new \Silex\Provider\SecurityServiceProvider(), array(
+            'security.firewalls' => array(
+                'admin' => array(
+                    'pattern' => '^/admin',
+                    'http' => true,
+                    'users' => array(),
+                ),
+            ),
+        ));
+
+        $app->get("/admin", function () {
+            return "SECURE!";
+        });
+
+        $request = Request::create("/admin");
+        $app->run($request);
+
+        $this->assertEmpty($app['monolog.handler']->getRecords(), "Expected no logging to occur");
     }
 
     protected function assertMatchingRecord($pattern, $level, $handler)