+++ tests/Silex/Tests/RouterTest.php
@@ -16,7 +16,6 @@ use Silex\Application;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * Router test cases.
@@ -97,7 +96,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-    * @expectedException Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+    * @expectedException \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
     */
     public function testMissingRoute()
     {
@@ -179,6 +178,18 @@ class RouterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('/foo/', $response->getTargetUrl());
     }
 
+    public function testHostSpecification()
+    {
+        if (!method_exists('Symfony\Component\Routing\Route', 'setHost')) {
+            $this->markTestSkipped('host() is only supported in the Symfony Routing 2.2+');
+        }
+
+        $route = new \Silex\Route();
+
+        $this->assertSame($route, $route->host('{locale}.example.com'));
+        $this->assertEquals('{locale}.example.com', $route->getHost());
+    }
+
     public function testRequireHttpRedirect()
     {
         $app = $this->createApplication();
@@ -205,6 +216,20 @@ class RouterTest extends \PHPUnit_Framework_TestCase
         $request = Request::create('http://example.com/secured');
         $response = $app->handle($request);
         $this->assertTrue($response->isRedirect('https://example.com/secured'));
+    }
+
+    public function testRequireHttpsRedirectIncludesQueryString()
+    {
+        $app = new Application();
+
+        $app->match('/secured', function () {
+            return 'secured content';
+        })
+        ->requireHttps();
+
+        $request = Request::create('http://example.com/secured?query=string');
+        $response = $app->handle($request);
+        $this->assertTrue($response->isRedirect('https://example.com/secured?query=string'));
     }
 
     public function testClassNameControllerSyntax()